+++ src/Clock.cpp
@@ -2,7 +2,7 @@
 #include "easylogging++.h"
 
 Clock::Clock(const std::shared_ptr<const Config> config)
- : _time(0.f)
+ : _time(0)
  , _config(config)
 {
   LOG(TRACE) << "Clock created.";
@@ -13,18 +13,18 @@ Clock::~Clock()
   LOG(TRACE) << "Clock created.";
 }
 
-float Clock::currentTime() const
+unsigned long Clock::currentTime() const
 {
   return _time;
 }
 
-void Clock::advanceTo(const float time)
+void Clock::advanceTo(const unsigned long time)
 {
   if (time <= _time) return;
   _time = time;
 }
 
-void Clock::advanceBy(const float amount)
+void Clock::advanceBy(const unsigned long amount)
 {
   if (amount <= 0) return;
   _time += amount;
